Fix column and index-size for MySQL only (#1185)
* Fix column and index-size for MySQL only
Conflicts:
	modules/saml/lib/SP/LogoutStore.php

+++ modules/saml/lib/SP/LogoutStore.php
@@ -2,6 +2,13 @@
 
 namespace SimpleSAML\Module\saml\SP;
 
+use PDO;
+use SAML2\XML\saml\NameID;
+use SimpleSAML\Logger;
+use SimpleSAML\Session;
+use SimpleSAML\Store;
+use SimpleSAML\Utils;
+
 /**
  * A directory over logout information.
  *
@@ -14,31 +21,84 @@ class LogoutStore
      * Create logout table in SQL, if it is missing.
      *
      * @param \SimpleSAML\Store\SQL $store  The datastore.
+     * @return void
      */
-    private static function createLogoutTable(\SimpleSAML\Store\SQL $store)
+    private static function createLogoutTable(Store\SQL $store)
     {
         $tableVer = $store->getTableVersion('saml_LogoutStore');
-        if ($tableVer === 3) {
+        if ($tableVer === 4) {
             return;
+        } elseif ($tableVer === 3) {
+            /**
+             * Table version 4 fixes the column type for the _expire column.
+             * We now use DATETIME instead of TIMESTAMP to support MSSQL.
+             */
+            switch ($store->driver) {
+                case 'pgsql':
+                    // This does not affect the NOT NULL constraint
+                    $update = ['ALTER TABLE '.$store->prefix.
+                        '_saml_LogoutStore ALTER COLUMN _expire DATETIME'];
+                    break;
+                case  'sqlite':
+                    /**
+                     * Because SQLite does not support field alterations, the approach is to:
+                     *     Create a new table without the proper column size
+                     *     Copy the current data to the new table
+                     *     Drop the old table
+                     *     Rename the new table correctly
+                     *     Read the index
+                     */
+                    $update = [
+                        'CREATE TABLE '.$store->prefix.'_saml_LogoutStore_new (_authSource VARCHAR(255) NOT NULL,'.
+                        '_nameId VARCHAR(40) NOT NULL, _sessionIndex VARCHAR(50) NOT NULL, _expire DATETIME NOT NULL,'.
+                        '_sessionId VARCHAR(50) NOT NULL, UNIQUE (_authSource, _nameID, _sessionIndex))',
+                        'INSERT INTO '.$store->prefix.'_saml_LogoutStore_new SELECT * FROM '.$store->prefix.'_saml_LogoutStore',
+                        'DROP TABLE '.$store->prefix.'_saml_LogoutStore',
+                        'ALTER TABLE '.$store->prefix.'_saml_LogoutStore_new RENAME TO '.$store->prefix.'_saml_LogoutStore',
+                        'CREATE INDEX '.$store->prefix.'_saml_LogoutStore_expire ON '.$store->prefix.'_saml_LogoutStore (_expire)',
+                        'CREATE INDEX '.$store->prefix.'_saml_LogoutStore_nameId ON '.$store->prefix.'_saml_LogoutStore (_authSource, _nameId)'
+                    ];
+                    break;
+                default:
+                    $update = ['ALTER TABLE '.$store->prefix.
+                        '_saml_LogoutStore MODIFY _expire DATETIME NOT NULL'];
+                    break;
+            }
+
+            try {
+                foreach ($update as $query) {
+                    $store->pdo->exec($query);
+                }
+            } catch (\Exception $e) {
+                Logger::warning('Database error: '.var_export($store->pdo->errorInfo(), true));
+                return;
+            }
+            $store->setTableVersion('saml_LogoutStore', 4);
+            return;
+
         } elseif ($tableVer === 2) {
-            // Drop old indexes
-            $query = 'ALTER TABLE '.$store->prefix.'_saml_LogoutStore DROP INDEX '.$store->prefix.'_saml_LogoutStore_nameId';
-            $store->pdo->exec($query);
-            $query = 'ALTER TABLE '.$store->prefix.'_saml_LogoutStore DROP INDEX _authSource';
-            $store->pdo->exec($query);
+            // TableVersion 3 fixes the indexes that were set to 255 in version 2; they cannot be larger than 191 on MySQL
 
-            // Create new indexes
-            $query = 'CREATE INDEX '.$store->prefix.'_saml_LogoutStore_nameId ON ';
-            $query .= $store->prefix.'_saml_LogoutStore (_authSource(191), _nameId)';
-            $store->pdo->exec($query);
+            if ($store->driver === 'mysql') {
+                // Drop old indexes
+                $query = 'ALTER TABLE '.$store->prefix.'_saml_LogoutStore DROP INDEX '.$store->prefix.'_saml_LogoutStore_nameId';
+                $store->pdo->exec($query);
+                $query = 'ALTER TABLE '.$store->prefix.'_saml_LogoutStore DROP INDEX _authSource';
+                $store->pdo->exec($query);
 
-            $query = 'ALTER TABLE '.$store->prefix.'_saml_LogoutStore ADD UNIQUE KEY (_authSource(191), _nameID, _sessionIndex)';
-            $store->pdo->exec($query);
+                // Create new indexes
+                $query = 'CREATE INDEX '.$store->prefix.'_saml_LogoutStore_nameId ON ';
+                $query .= $store->prefix.'_saml_LogoutStore (_authSource(191), _nameId)';
+                $store->pdo->exec($query);
+
+                $query = 'ALTER TABLE '.$store->prefix.'_saml_LogoutStore ADD UNIQUE KEY (_authSource(191), _nameID, _sessionIndex)';
+                $store->pdo->exec($query);
+            }
 
             $store->setTableVersion('saml_LogoutStore', 3);
             return;
         } elseif ($tableVer === 1) {
-            // TableVersion 2 increased the column size to 255 which is the maximum length of a FQDN
+            // TableVersion 2 increased the column size to 255 (191 for mysql) which is the maximum length of a FQDN
             switch ($store->driver) {
                 case 'pgsql':
                     // This does not affect the NOT NULL constraint
@@ -47,7 +107,6 @@ class LogoutStore
                     break;
                 case  'sqlite':
                     /**
-                     * TableVersion 2 increased the column size to 255 which is the maximum length of a FQDN
                      * Because SQLite does not support field alterations, the approach is to:
                      *     Create a new table without the proper column size
                      *     Copy the current data to the new table
@@ -66,6 +125,10 @@ class LogoutStore
                         'CREATE INDEX '.$store->prefix.'_saml_LogoutStore_nameId ON '.$store->prefix.'_saml_LogoutStore (_authSource, _nameId)'
                     ];
                     break;
+                case 'mysql':
+                    $update = ['ALTER TABLE '.$store->prefix.
+                        '_saml_LogoutStore MODIFY _authSource VARCHAR(191) NOT NULL'];
+                    break;
                 default:
                     $update = ['ALTER TABLE '.$store->prefix.
                         '_saml_LogoutStore MODIFY _authSource VARCHAR(255) NOT NULL'];
@@ -77,7 +140,7 @@ class LogoutStore
                     $store->pdo->exec($query);
                 }
             } catch (\Exception $e) {
-                \SimpleSAML\Logger::warning($store->pdo->errorInfo());
+                Logger::warning('Database error: '.var_export($store->pdo->errorInfo(), true));
                 return;
             }
             $store->setTableVersion('saml_LogoutStore', 2);
@@ -85,12 +148,12 @@ class LogoutStore
         }
 
         $query = 'CREATE TABLE '.$store->prefix.'_saml_LogoutStore (
-            _authSource VARCHAR(255) NOT NULL,
+            _authSource VARCHAR('.($store->driver === 'mysql' ? '191' : '255').') NOT NULL,
             _nameId VARCHAR(40) NOT NULL,
             _sessionIndex VARCHAR(50) NOT NULL,
-            _expire TIMESTAMP NOT NULL,
+            _expire DATETIME NOT NULL,
             _sessionId VARCHAR(50) NOT NULL,
-            UNIQUE (_authSource(191), _nameID, _sessionIndex)
+            UNIQUE (_authSource'.($store->driver === 'mysql' ? '(191)' : '').', _nameID, _sessionIndex)
         )';
         $store->pdo->exec($query);
 
@@ -99,10 +162,10 @@ class LogoutStore
         $store->pdo->exec($query);
 
         $query = 'CREATE INDEX '.$store->prefix.'_saml_LogoutStore_nameId ON ';
-        $query .= $store->prefix.'_saml_LogoutStore (_authSource(191), _nameId)';
+        $query .= $store->prefix.'_saml_LogoutStore (_authSource'.($store->driver === 'mysql' ? '(191)' : '').', _nameId)';
         $store->pdo->exec($query);
 
-        $store->setTableVersion('saml_LogoutStore', 3);
+        $store->setTableVersion('saml_LogoutStore', 4);
     }
 
 
@@ -110,10 +173,11 @@ class LogoutStore
      * Clean the logout table of expired entries.
      *
      * @param \SimpleSAML\Store\SQL $store  The datastore.
+     * @return void
      */
-    private static function cleanLogoutStore(\SimpleSAML\Store\SQL $store)
+    private static function cleanLogoutStore(Store\SQL $store)
     {
-        \SimpleSAML\Logger::debug('saml.LogoutStore: Cleaning logout store.');
+        Logger::debug('saml.LogoutStore: Cleaning logout store.');
 
         $query = 'DELETE FROM '.$store->prefix.'_saml_LogoutStore WHERE _expire < :now';
         $params = ['now' => gmdate('Y-m-d H:i:s')];
@@ -130,9 +194,12 @@ class LogoutStore
      * @param string $authId  The authsource ID.
      * @param string $nameId  The hash of the users NameID.
      * @param string $sessionIndex  The SessionIndex of the user.
+     * @param int $expire
+     * @param string $sessionId
+     * @return void
      */
     private static function addSessionSQL(
-        \SimpleSAML\Store\SQL $store,
+        Store\SQL $store,
         $authId,
         $nameId,
         $sessionIndex,
@@ -142,8 +209,8 @@ class LogoutStore
         assert(is_string($authId));
         assert(is_string($nameId));
         assert(is_string($sessionIndex));
-        assert(is_string($sessionId));
         assert(is_int($expire));
+        assert(is_string($sessionId));
 
         self::createLogoutTable($store);
 
@@ -174,7 +241,7 @@ class LogoutStore
      * @param string $nameId  The hash of the users NameID.
      * @return array  Associative array of SessionIndex =>  SessionId.
      */
-    private static function getSessionsSQL(\SimpleSAML\Store\SQL $store, $authId, $nameId)
+    private static function getSessionsSQL(Store\SQL $store, $authId, $nameId)
     {
         assert(is_string($authId));
         assert(is_string($nameId));
@@ -194,7 +261,7 @@ class LogoutStore
         $query->execute($params);
 
         $res = [];
-        while (($row = $query->fetch(\PDO::FETCH_ASSOC)) !== false) {
+        while (($row = $query->fetch(PDO::FETCH_ASSOC)) !== false) {
             $res[$row['_sessionindex']] = $row['_sessionid'];
         }
 
@@ -211,7 +278,7 @@ class LogoutStore
      * @param array $sessionIndexes  The session indexes.
      * @return array  Associative array of SessionIndex =>  SessionId.
      */
-    private static function getSessionsStore(\SimpleSAML\Store $store, $authId, $nameId, array $sessionIndexes)
+    private static function getSessionsStore(Store $store, $authId, $nameId, array $sessionIndexes)
     {
         assert(is_string($authId));
         assert(is_string($nameId));
@@ -242,6 +309,8 @@ class LogoutStore
      * @param string $authId  The authsource ID.
      * @param \SAML2\XML\saml\NameID $nameId The NameID of the user.
      * @param string|null $sessionIndex  The SessionIndex of the user.
+     * @param int $expire
+     * @return void
      */
     public static function addSession($authId, $nameId, $sessionIndex, $expire)
     {
@@ -255,10 +324,10 @@ class LogoutStore
              * it supports SLO, but we don't want an LogoutRequest with a specific
              * SessionIndex to match this session. We therefore generate our own session index.
              */
-            $sessionIndex = \SimpleSAML\Utils\Random::generateID();
+            $sessionIndex = Utils\Random::generateID();
         }
 
-        $store = \SimpleSAML\Store::getInstance();
+        $store = Store::getInstance();
         if ($store === false) {
             // We don't have a datastore.
             return;
@@ -267,7 +336,7 @@ class LogoutStore
         // serialize and anonymize the NameID
         // TODO: remove this conditional statement
         if (is_array($nameId)) {
-            $nameId = \SAML2\XML\saml\NameID::fromArray($nameId);
+            $nameId = NameID::fromArray($nameId);
         }
         $strNameId = serialize($nameId);
         $strNameId = sha1($strNameId);
@@ -277,10 +346,10 @@ class LogoutStore
             $sessionIndex = sha1($sessionIndex);
         }
 
-        $session = \SimpleSAML\Session::getSessionFromRequest();
+        $session = Session::getSessionFromRequest();
         $sessionId = $session->getSessionId();
 
-        if ($store instanceof \SimpleSAML\Store\SQL) {
+        if ($store instanceof Store\SQL) {
             self::addSessionSQL($store, $authId, $strNameId, $sessionIndex, $expire, $sessionId);
         } else {
             $store->set('saml.LogoutStore', $strNameId.':'.$sessionIndex, $sessionId, $expire);
@@ -294,13 +363,13 @@ class LogoutStore
      * @param string $authId  The authsource ID.
      * @param \SAML2\XML\saml\NameID $nameId The NameID of the user.
      * @param array $sessionIndexes  The SessionIndexes we should log out of. Logs out of all if this is empty.
-     * @returns int|false  Number of sessions logged out, or FALSE if not supported.
+     * @return int|false  Number of sessions logged out, or FALSE if not supported.
      */
     public static function logoutSessions($authId, $nameId, array $sessionIndexes)
     {
         assert(is_string($authId));
 
-        $store = \SimpleSAML\Store::getInstance();
+        $store = Store::getInstance();
         if ($store === false) {
             // We don't have a datastore
             return false;
@@ -309,7 +378,7 @@ class LogoutStore
         // serialize and anonymize the NameID
         // TODO: remove this conditional statement
         if (is_array($nameId)) {
-            $nameId = \SAML2\XML\saml\NameID::fromArray($nameId);
+            $nameId = NameID::fromArray($nameId);
         }
         $strNameId = serialize($nameId);
         $strNameId = sha1($strNameId);
@@ -325,13 +394,13 @@ class LogoutStore
         // Remove reference
         unset($sessionIndex);
 
-        if ($store instanceof \SimpleSAML\Store\SQL) {
+        if ($store instanceof Store\SQL) {
             $sessions = self::getSessionsSQL($store, $authId, $strNameId);
         } elseif (empty($sessionIndexes)) {
             // We cannot fetch all sessions without a SQL store
             return false;
         } else {
-            /** @var \SimpleSAML\Store $sessions At this point the store cannot be false */
+            /** @var array $sessions At this point the store cannot be false */
             $sessions = self::getSessionsStore($store, $authId, $strNameId, $sessionIndexes);
         }
 
@@ -342,26 +411,26 @@ class LogoutStore
         $numLoggedOut = 0;
         foreach ($sessionIndexes as $sessionIndex) {
             if (!isset($sessions[$sessionIndex])) {
-                \SimpleSAML\Logger::info('saml.LogoutStore: Logout requested for unknown SessionIndex.');
+                Logger::info('saml.LogoutStore: Logout requested for unknown SessionIndex.');
                 continue;
             }
 
             $sessionId = $sessions[$sessionIndex];
 
-            $session = \SimpleSAML\Session::getSession($sessionId);
+            $session = Session::getSession($sessionId);
             if ($session === null) {
-                \SimpleSAML\Logger::info('saml.LogoutStore: Skipping logout of missing session.');
+                Logger::info('saml.LogoutStore: Skipping logout of missing session.');
                 continue;
             }
 
             if (!$session->isValid($authId)) {
-                \SimpleSAML\Logger::info(
+                Logger::info(
                     'saml.LogoutStore: Skipping logout of session because it isn\'t authenticated.'
                 );
                 continue;
             }
 
-            \SimpleSAML\Logger::info(
+            Logger::info(
                 'saml.LogoutStore: Logging out of session with trackId ['.$session->getTrackID().'].'
             );
             $session->doLogout($authId);
